pyproject.toml
[project]
name = "data-mastor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
# CHORE: make sure renovate does not try to auto-update python version
requires-python = ">=3.13,<3.14"
dependencies = [
    "deepdiff>=8.6.1",
    "pandas>=2.3.3",
    "psutil>=7.1.1",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "scrapy>=2.13.3",
    "sqlalchemy>=2.0.44",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "pre-commit>=4.4.0",
    "pytest>=9.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.4",
    "types-psutil>=7.0.0.20251001",
    "types-pyyaml>=6.0.12.20250915",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
dm = "data_mastor.cli:app"

# TOOLS
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov-config pyproject.toml --cov-report=term --cov-report=html --cov-report=xml"

[tool.coverage.run]
source_pkgs = ["data_mastor"]

[tool.coverage.report]
omit = ["__init__.py"]

[tool.ruff]
[tool.ruff.lint]
per-file-ignores = { "tests/scraper/conftest.py" = ["F401"] }

[tool.mypy]
packages=["data_mastor"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
check_untyped_defs = true

